+++ piezas/bishop.py
@@ -1,0 +1,4 @@
+from piezas.piece import Piece
+
+class Bishop(Piece):
+    ...