+++ piezas/piece.py
@@ -1,0 +1,3 @@
+class Piece:    #Clase Padre
+    def __init__(self, color):
+        self.__color__ = color