+++ test/test_board.py
@@ -1,6 +1,8 @@
 import unittest
 from game.board import Board
-from piece.rook import Rook
+from piezas.rook import Rook
+from piezas.queen import Queen
+from piezas.horse import Horse
 
 class TestBoard(unittest.TestCase):
 
@@ -10,16 +12,31 @@ class TestBoard(unittest.TestCase):
         self.assertEqual(len(board.grid[0]),8,)
 
         # Verifica la posición de las torres negras
-        self.assertIsInstance(board._Board__positions__[0][0], Rook)
-        self.assertEqual(board._Board__positions__[0][0].color, "BLACK")
-        self.assertIsInstance(board._Board__positions__[0][7], Rook)
-        self.assertEqual(board._Board__positions__[0][7].color, "BLACK")
-
+        self.assertIsInstance(board.get_piece[0][0], Rook) #BLACK
+        self.assertIsInstance(board.get_piece[0][7], Rook) #BLACK
+        
         # Verifica la posición de las torres blancas
-        self.assertIsInstance(board._Board__positions__[7][0], Rook)
-        self.assertEqual(board._Board__positions__[7][0].color, "WHITE")
-        self.assertIsInstance(board._Board__positions__[7][7], Rook)
-        self.assertEqual(board._Board__positions__[7][7].color, "WHITE")
+        self.assertIsInstance(board.get_piece[7][0], Rook) #WHITE
+        self.assertIsInstance(board.get_piece[7][7], Rook) #WHITE
+
+        # Verifica la posición de la reina negra
+        self.assertIsInstance(board.get_piece[0][3], Queen) #BLACK
+      
+        # Verifica la posición de la reina blancas
+        self.assertIsInstance(board.get_piece[7][3], Queen)#WHITE
+        
+
+        # Verifica la posición de los caballos negros
+        self.assertIsInstance(board.get_piece[0][2], Horse)#BLACK
+        self.assertIsInstance(board.get_piece[0][5], Horse) #BLACK
+   
+
+        # Verifica la posición de los caballos blancos
+        self.assertIsInstance(board.get_piece[7][2], Horse)#WHITE
+        self.assertIsInstance(board.get_piece[7][5], Horse) #WHITE
+   
+
+
 
 if __name__ == '__main__':
     unittest.main()